Changed the geometry initialisation procedure (sorting)
Before this change, GeometryCore asked ChannelMapAlg to "Initialize()", and ChannelMapAlg would both sort the geometry and then initialise the mapping.
Now GeometryCore asks ChannelMapAlg for the sorter, then uses it to sort the geometry, then asks ChannelMapAlg to Initialize() itself
(of course this initialisation does not include sorting any more).